Moved the URI class to the Utils/Http directory.
